+++ latex2mathml/aggregator.py
@@ -3,32 +3,35 @@ from tokenizer import tokenize
 
 __author__ = 'Ronie Martinez'
 
+matrix_commands = (r'\matrix', r'\matrix*',
+                   r'\pmatrix', r'\pmatrix*',
+                   r'\bmatrix', r'\bmatrix*',
+                   r'\Bmatrix', r'\Bmatrix*',
+                   r'\vmatrix', r'\vmatrix*',
+                   r'\Vmatrix', r'\Vmatrix*')
+
 
 def aggregate(latex):
     aggregation = []
     subgroups = [aggregation]
     insert_before_last_item = False
     environment = None
+    has_negative_sign = False
     for token in tokenize(latex):
-        if token == r'\matrix':
-            environment = r'\matrix'
-        if token in '{([':
+        if token in matrix_commands:
+            environment = token
+            _insert_before_last_item(insert_before_last_item, token, subgroups)
+        elif token in '{([':
             try:
                 a = subgroups[-1][-1]
                 if a != r'\left':
                     raise IndexError
             except IndexError:
                 n = []
-                if insert_before_last_item:
-                    subgroups[-1].insert(-1, n)
-                    insert_before_last_item = False
-                else:
-                    subgroups[-1].append(n)
+                _insert_before_last_item(insert_before_last_item, n, subgroups)
                 subgroups.append(n)
-            if environment and environment == r'\matrix':
-                n = []
-                subgroups[-1].append(n)
-                subgroups.append(n)
+            if environment and environment in matrix_commands:
+                _add_new_subgroup(subgroups)
             elif token == '[' and subgroups[-2][-2] == r'\sqrt':
                 subgroups[-2][-2] = r'\root'  # change name from \sqrt to \root - not a latex command!
             elif token != '{':
@@ -58,17 +61,34 @@ def aggregate(latex):
                     subgroups[-1].insert(-1, token)
             except IndexError:
                 subgroups[-1].insert(-1, token)
-        elif token == '&' and environment and environment == r'\matrix':
-            pass
-        elif token == r'\\' and environment and environment == r'\matrix':
+        elif token == '-' and environment and environment in matrix_commands:
+            _add_new_subgroup(subgroups)
+            _insert_before_last_item(insert_before_last_item, token, subgroups)
+            has_negative_sign = True
+        elif token == '&' and environment and environment in matrix_commands:
+            if has_negative_sign:
+                subgroups.pop()
+                has_negative_sign = False
+        elif (token == r'\\' or token == r'\cr') and environment and environment in matrix_commands:
+            if has_negative_sign:
+                subgroups.pop()
+                has_negative_sign = False
             subgroups.pop()
-            n = []
-            subgroups[-1].append(n)
-            subgroups.append(n)
+            _add_new_subgroup(subgroups)
         else:
-            if insert_before_last_item:
-                subgroups[-1].insert(-1, token)
-                insert_before_last_item = False
-            else:
-                subgroups[-1].append(token)
+            _insert_before_last_item(insert_before_last_item, token, subgroups)
     return aggregation
+
+
+def _insert_before_last_item(insert_before_last_item, n, subgroups):
+    if insert_before_last_item:
+        subgroups[-1].insert(-1, n)
+        insert_before_last_item = False
+    else:
+        subgroups[-1].append(n)
+
+
+def _add_new_subgroup(subgroups):
+    n = []
+    subgroups[-1].append(n)
+    subgroups.append(n)