+++ tests/tokenizer_test.py
@@ -53,5 +53,13 @@ class TokenizerTest(unittest.TestCase):
         self.assertListEqual([r'\matrix', '{', 'a', '&', 'b', r'\\', 'c', '&', 'd', '}'],
                              list(latex2mathml.tokenize(r'\begin{matrix}a & b \\ c & d \end{matrix}')))
 
+    def test_matrix_with_alignment(self):
+        self.assertListEqual([r'\matrix*', 'r', '{', 'a', '&', 'b', r'\\', 'c', '&', 'd', '}'],
+                             list(latex2mathml.tokenize(r'\begin{matrix*}[r]a & b \\ c & d \end{matrix*}')))
+
+    def test_matrix_with_negative_sign(self):
+        self.assertListEqual([r'\matrix', '{', '-', 'a', '&', 'b', r'\\', 'c', '&', 'd', '}'],
+                             list(latex2mathml.tokenize(r'\begin{matrix}-a & b \\ c & d \end{matrix}')))
+
 if __name__ == '__main__':
     unittest.main()